mediasense-backend/requirements.txt
# 基础依赖
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.2
billiard==4.2.1
celery==5.3.6
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Django==5.1.4
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-timezone-field==7.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
elasticsearch==7.17.9
frozenlist==1.5.0
gunicorn==21.2.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
mysqlclient==2.2.0
openai==0.28.0
packaging==24.2
prompt_toolkit==3.0.48
propcache==0.2.1
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
redis==5.0.1
requests==2.31.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3

# Excel处理
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1
xlwt==1.3.0

# 代码质量工具
black==24.3.0
isort==5.13.2
flake8==7.0.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
flake8-quotes==3.4.0
flake8-bugbear==24.2.6
pylint==3.1.0
pylint-django==2.5.5
mypy==1.9.0
radon==6.0.1
bandit==1.7.8
safety==3.0.1
pre-commit==3.6.2

# 类型检查支持
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# 测试工具
pytest==8.1.1
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-html==4.1.1
factory-boy==3.3.0
faker==24.3.0

# 性能分析工具
django-silk==5.1.0
django-cacheops==7.0.2
django-querycount==0.8.3

# 文档工具
drf-yasg==1.21.7
sphinx==7.2.6
mkdocs==1.5.3
mkdocs-material==9.5.13
